% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.setts.attr.R
\name{add.setts.attr}
\alias{add.setts.attr}
\title{add.setts.attr}
\usage{
add.setts.attr(simulation, settings)
}
\arguments{
\item{simulation}{simulation from sim_dauer, sim_dauer_unbal, or sim_dauer_biased
requires a 'settings' list in the environment}

\item{settings}{list of settings from simluations}
}
\description{
add attributes to simualtion based on input settigns
}
\examples{
simulation \%<>\% add.setts.attr
}
