% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alpha.R
\name{get_alpha}
\alias{get_alpha}
\title{summarize results of model simulations to get proportion p <0.05
 doesn't work if too few simulations and no p < 0.05 for one of the models}
\usage{
get_alpha(df, lmm)
}
\arguments{
\item{df}{simulation data frame from sim_dauer, sim_dauer_unbal, or sim_dauer_biased. Pre-processed
using add.setts.attr to add attributes}

\item{nsim}{}
}
\description{
summarize results of model simulations to get proportion p <0.05
 doesn't work if too few simulations and no p < 0.05 for one of the models
}
\examples{
balanced.mod \%<>\% get_alpha()
}
