% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par.replicate.R
\name{par.replicate}
\alias{par.replicate}
\title{par.replicate}
\usage{
par.replicate(cl, n, expr, simplify = FALSE)
}
\arguments{
\item{cl}{cluster connections from parallel::makeCluster}

\item{n}{number of replicates, or number of simulations}

\item{expr}{expression or function to use in replicate}

\item{simplify}{wrapper to simplify within parSapply, usually FALSE}
}
\description{
wrapper fo parSapply to mimic replicate in base R for parallel
}
\examples{
library(parallel) 
cl <- makeCluster(2) 
clusterEvalQ(cl,  { library(MASS)
library(magrittr)
library(dplyr)
library(lme4)
library(rstan)
library(rstanarm)
})
clusterExport(cl=cl, varlist=c("sim_dauer_3G_stan", "settings"))
simulation <- do.call( rbind, par.replicate(cl,n=1000, 
sim_dauer_3G_stan(c(settings,
do.plot = FALSE, 
do.stan = TRUE)), 
simplify=FALSE ))

}
