% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dauer.R
\name{sim_dauer}
\alias{sim_dauer}
\title{sim_dauer}
\usage{
sim_dauer(settings)
}
\arguments{
\item{settings}{input list of settings for the simulation. Must be a list.
settings$I    #population control intercept (in logit)
settings$nP  # number of plates
settings$nD  # number of days
settings$sP  # plate to plate sd
settings$sD  # day to day sd
settings$sG  # genotype sd due to culture history (logit)
settings$k   # number animal per plate)
settings$A    # population A intercept (expt)
settings$B    # pop B intercept}
}
\description{
simulate balanced dauer assay data using 3 groups of strains
I, A and B. Simulation uses day-to-day and plate-to-plate variance. 
Can be used to simulate any binomial data with hierarchical clusters.
}
\examples{
settings <- list(settings <- list(
I = 0 #population control intercept (in logit). 0 = p(0.5)
,nP = 6 # number of plates
,nD = 3 # number of days
,sP = 0.1 # plate to plate variance (0.3)
,sD = 0.5 # day to day variance (0.2)
,sG = 0.5 # genotype variance due to culture history (logit) (0.2)
,k = 60 # number animal per plate)
,A = 0 # population A intercept (expt - genotype2)
,B = 0 #pop B intercept (expt - genotype2)
))


sim_dauer(settings = c(settings, do.plot = TRUE))

simulation <- sim_dauer(settings = c(settings, do.plot = FALSE, do.stan = TRUE))
}
